refactor(cli): scinder _commun, et tenir les 400 lignes
Le découpage précédent laissait _commun.py à 701 lignes, au-dessus du plafond
que l'issue fixe. Il est scindé en trois, selon le MOMENT où chaque helper
intervient dans une commande : la résolution du contexte reste dans _commun,
_amorcage porte ce qui encadre toute invocation (le callback global, la version,
l'avis de mise à jour), et _validation porte le verdict.

Aucun fichier de cli/ ne dépasse désormais 380 lignes. Le corps des fonctions
est déplacé verbatim ; seuls les en-têtes sont neufs.

Deux patchs de test_contrat_honore.py étaient inertes, et le découpage l'a
révélé : ils posaient l'attribut sur le paquet quand progression.submit appelle
sa propre liaison. Ces tests jouaient donc sur le vrai résultat, 0/100, et non
sur les 40/100 qu'ils annoncent pour éprouver un seuil à 70. L'adresse est
corrigée, avec la raison en commentaire, et les deux passent toujours : rien ne
se cachait derrière, submit gère bien son seuil. Un test qui croit mesurer et ne
mesure rien reste plus dangereux qu'un test absent, parce qu'il rassure.

Vérifié : 705 tests unitaires, 18 e2e, ruff, mypy strict sur 78 fichiers, et
l'ordre de dsoxlab --help identique à la référence, comparé par diff.
